Add fabric task for local development of recipes
Run 'fab run'. It will make a dev copy of recipes.json in which all JS will be served on http://localhost:8000/ instead of whatever it was set at. Note that this means that if some recipes are not actually in the repo, they will not work.

The port 8000 can be changed on the command line by running 'fab run:port=8888' by note that the dev extension will still be expecting the recipes to be served from port 8000, so you would need to change that as well.

The development process would then be:
- enable the dev extension (which means, you probably have to run the remoats-server locally as well)
- pull recipes repo
- fab run in "recipes" repo
- go to website you want to work on
- modify recipe file from inside the "recipes" repo
- refresh every time
- once done, you can stop the recipes server, commit and push.

If you change the cookbook (recipes.json), you might have to reload the extension to accept new changes. (it depends on when the extension loads that file)